Rename pedadocument -> pedagogicaldocument
